[WEB] use-effect-studies - update: creating components for the useEffect studies
