+++ src/App.jsx
@@ -6,64 +6,7 @@ import NameList from './components/NameList'
 import ReducerNameList from './components/ReducerNameList';
 import UserForm from './components/UserForm';
 import MemoCallbackComponent from './components/MemoCallbackComponent';
-import { useState, useEffect } from 'react';
-
-function Stopwatch() {
-  const [time, setTime] = useState(0)
-
-  useEffect(() => {
-    const intervalID = setInterval(() => {
-      setTime(oldTime => oldTime + 1)
-    }, 1000)
-    
-    return () => clearInterval(intervalID)
-  }, [])
-
-  return (
-    <div>
-      <p>Time: {time}</p>
-    </div>
-  )
-}
-
-function UseEffectComponent() {
-
-  const [names, setNames] = useState([])
-  
-  useEffect(() => {
-    fetch("/names.json")
-    .then((res) => res.json())
-    .then((data) => setNames(data))
-    .catch((e) => console.error('Error Message (catch): ', e))
-  }, [])
-  
-  const [selectedName, setSelectedName] = useState(null)
-  const [selectedNameDetails, setSelectedNameDetails] = useState(null)
-
-  const handleSelectedNameChange = (name) => {
-    fetch(`/${name}.json`)
-      .then((res) => res.json())
-      .then((data) => setSelectedNameDetails(data))
-      .catch((e) => console.error('Error Message (catch): ', e))
-  }
-  return (
-    <div>
-      <Stopwatch/>
-      <div>
-        {names.map((name, key) => (
-          <button 
-            key={key}
-            onClick={() => handleSelectedNameChange(name)}
-          >
-            {name}
-          </button>
-        ))}
-      </div>
-      <div>{JSON.stringify(selectedName)}</div>
-      <div>Selected Name Details: {JSON.stringify(selectedNameDetails)}</div>
-    </div>
-  )
-}
+import UseEffectComponent from './components/UseEffectComponent';
 
 function App() {
